dataset_generation/requirements.txt
requests
svglib
reportlab
python-dotenv